Add a log_hex parameter to tm_parse.
This is passed on to the constructors for the various TM subclasses. It
defaults to True, so shouldn't change any current behaviour. If set to
False, it prevents the *_LOG_FH writes that the TM constructors do. With
this change, the TM subclasses can be used for reading in previously
logged data.