Matt requested a change to CSV output
Conversion of science data to CSV should match previous output generated by
Dave's LabView scripts. I've added a "default_fields_per_type" to the
EGSEDumpDecoder class. It's not used directly by the class, but decoders
can use this as a field list handed to the csv() and csvheader() methods
of the underlying classes.

The __main__ section of egse_dump_decoder.py has an example of how this
might be used.