Fixes found by ruff check.
Some were in find_dac_offset - I ran ruff check in lieu of actually being
able to run the code. While I was at it, I fixed most of the other items
that ruff reported. There are a few "variable assigned but unused" warnings
remaining. I didn't "fix" those because it's not clear what the appropriate
course of action would be.