Add piecewise_linear function to abu_sequences.py
This will be used for estimating a dark offset target in the "bathtub"
range. The EB doesn't use floating point, so while we have an exponential
formula relating bathtub output offsets to temperature, we can't use the
formula. Instead, we have a table of points and we'll use interpolation to
get a close value.