Fix bug in DAC offset estimation.
When we start the estimation, we get a science packet to measure
HT_SINK_TEMP so we can estimate a target for binary chop. But if
the current DAC offset is causing underflow, this temperature
reading will be wrong, affecting the target estimation.

So we need to set DAC offsets to "a low value" before asking for
the temperature.

I've also been told that a DAC offset of zero is inadvisable,
and that using 1 is a better starting option.